Added SQL file
Here is what I've done so far for our SQL database. In the file I have written code to create two tables, both the clinic employee table and patient table. If you need any references as to what the schema should contain then please look at the ERD diagram we created for check-in #1 which you should be able to find in the images folder of this repo.