Update medicalClinicSQL.sql
Edited doctor and specialty, making specialty work properly as a multivalued attribute. Added approval table to track patient's approvals for specialists, added trigger a trigger that should check if a patient has specialist approval before an appointment is made, if there is no specialist approval it triggers a function that should tell the end user that they need a specific specialist approval to schedule a meeting with that doctor. Added Cascade option to all FOREIGN keys